move check function to collection.ts
